Refactor: Home 에서 Main으로 파일명 변경 및 분리되어있던 컴포넌트 MainPage와 합치기
fix #5
